requirements.txt
fastapi[all]
uvicorn
transformers
torch
pillow
requests